:recycle: cleanup SQL
